Test2.in to assign LENS PU via env
this file is a new implementation of getting the current states of
agents.  before the agent that has been picked to be updated will write
an agent.ex file, and the influencing agent will write an infl.ex file.
However, according to Mark, the way we want diffusion to work, using
an agent.ex file does not work.  Instead the .in file has been updated
to set the agent state by reading in an environment variable, p0-p4,
n0-n4